+++ src/main/java/open/commons/spring/web/validation/ValidationTarget.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2019 Park Jun-Hong_(parkjunhong77/google/com)
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ *
+ * This file is generated under this project, "open-commons-spring-web".
+ *
+ * Date  : 2019. 10. 15. 오후 3:06:02
+ *
+ * Author: Park_Jun_Hong_(fafanmama_at_naver_com)
+ * 
+ */
+
+package open.commons.spring.web.validation;
+
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.ElementType.PARAMETER;
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import javax.validation.Validation;
+
+/**
+ * {@link Validation} 대상임을 정의하는 어노테이션.
+ * 
+ * @since 2019. 10. 15.
+ * @version
+ * @author Park_Jun_Hong_(fafanmama_at_naver_com)
+ * 
+ * @see {@link Validation}
+ */
+@Documented
+@Retention(RUNTIME)
+@Target({ TYPE, FIELD, METHOD, PARAMETER })
+public @interface ValidationTarget {
+
+}